services/vector_search/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sentence-transformers==5.1.0
faiss-cpu==1.11.0.post1
numpy==2.3.2
pydantic==2.11.7